Use Hyphen-Minus instead of En Dash
As the Codeberg web UI informed me, these two characters can be confused
with one another.  And indeed, I did not notice that a non-standard dash
was being used here.  Tooling exists which extracts years from copyright
lines, and those expect hyphens.  Since the two characters look the same
(or nearly the same) with most monospace fonts, the only noticeable
effect of switching to the conventional character is, that the tooling
works again.